requirements.txt
Django>=1.11
amqp
billiard
boto3
botocore
celery
certifi
chardet
django-appconf
django-apps
django-colorful
django-cors-headers
django-http-proxy
django-imagekit
django-raster
django-wms
docutils
idna
jmespath
kombu
numpy
pilkit
Pillow
psycopg2
#pygdal
#https://gis.stackexchange.com/questions/28966/python-gdal-package-missing-header-file-when-installing-via-pip
GDAL==2.1.0
pyparsing
python-dateutil
python-slugify
pytz
requests
s3transfer
six
Unidecode
urllib3
vine

